+++ src/external_plugins/term_buffer/internal_term_buffer.cpp
@@ -518,17 +518,17 @@ void __InternalTermBuffer::ClearHistoryLinesData() {
     m_VisRowScrollRegionBegin = HasScrollRegion() ? m_ScrollRegionBegin : 0;
 }
 
-uint16_t __InternalTermBuffer::GetMode() const {
-    return (uint16_t)m_Mode.to_ulong();
+uint32_t __InternalTermBuffer::GetMode() const {
+    return (uint32_t)m_Mode.to_ulong();
 }
 
-void __InternalTermBuffer::SetMode(uint16_t m) {
+void __InternalTermBuffer::SetMode(uint32_t m) {
     m_Mode = std::bitset<16>(m);
 }
 
-void __InternalTermBuffer::AddMode(uint16_t m) {
+void __InternalTermBuffer::AddMode(uint32_t m) {
     m_Mode.set(m);
 }
-void __InternalTermBuffer::RemoveMode(uint16_t m) {
+void __InternalTermBuffer::RemoveMode(uint32_t m) {
     m_Mode.reset(m);
 }
